+++ src/Hellion.Cluster/ISC/InterConnector.cs
@@ -31,9 +31,7 @@ namespace Hellion.Cluster.ISC
         {
             var packetHeaderNumber = packet.Read<int>();
             var packetHeader = (InterHeaders)packetHeaderNumber;
-
-            Log.Debug("Recieved: {0}", packetHeader);
-
+            
             switch (packetHeader)
             {
                 case InterHeaders.CanAuthticate: this.Authenticate(); break;